+++ sites/partners/src/listings/PaperListingDetails/sections/DetailLeasingAgent.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from "react"
 import { t, GridSection, ViewItem, GridCell } from "@bloom-housing/ui-components"
 import { ListingContext } from "../../ListingContext"
-import { getDetailFieldString } from "./helpers"
+import { getDetailFieldString, getDetailAddress } from "./helpers"
 
 const DetailLeasingAgent = () => {
   const listing = useContext(ListingContext)
@@ -42,6 +42,11 @@ const DetailLeasingAgent = () => {
           </ViewItem>
         </GridCell>
       </GridSection>
+      {getDetailAddress(
+        listing.leasingAgentAddress,
+        "leasingAgentAddress",
+        t("listings.leasingAgentAddress")
+      )}
     </GridSection>
   )
 }