+++ sites/public/pages/listings.tsx
@@ -1,6 +1,9 @@
+import React, { useEffect, useContext } from "react"
 import Head from "next/head"
-import { ListingsGroup, PageHeader, t } from "@bloom-housing/ui-components"
+import { AuthContext, ListingsGroup, PageHeader, t } from "@bloom-housing/ui-components"
 import { Listing, ListingStatus } from "@bloom-housing/backend-core/types"
+import { ListingList, pushGtmEvent } from "@bloom-housing/shared-helpers"
+import { UserStatus } from "../lib/constants"
 import Layout from "../layouts/application"
 import { MetaTags } from "../src/MetaTags"
 import { getListings } from "../lib/helpers"
@@ -37,9 +40,20 @@ const closedListings = (listings) => {
 }
 
 export default function ListingsPage(props: ListingsProps) {
+  const { profile } = useContext(AuthContext)
   const pageTitle = `${t("pageTitle.rent")} - ${t("nav.siteTitle")}`
   const metaDescription = t("pageDescription.welcome", { regionName: t("region.name") })
   const metaImage = "" // TODO: replace with hero image
+
+  useEffect(() => {
+    pushGtmEvent<ListingList>({
+      event: "pageView",
+      pageTitle: "Rent Affordable Housing - Housing Portal",
+      status: profile ? UserStatus.LoggedIn : UserStatus.NotLoggedIn,
+      numberOfListings: props.openListings.length,
+      listingIds: props.openListings.map((listing) => listing.id),
+    })
+  }, [profile, props.openListings])
 
   return (
     <Layout>