+++ sites/partners/CHANGELOG.md
@@ -3,6 +3,170 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.1-alpha.44](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.43...@bloom-housing/partners@4.0.1-alpha.44) (2022-02-01)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.43](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.42...@bloom-housing/partners@4.0.1-alpha.43) (2022-02-01)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.42](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.41...@bloom-housing/partners@4.0.1-alpha.42) (2022-02-01)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.41](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.40...@bloom-housing/partners@4.0.1-alpha.41) (2022-02-01)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.40](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.39...@bloom-housing/partners@4.0.1-alpha.40) (2022-02-01)
+
+
+### Features
+
+* partners terms page ([#2440](https://github.com/bloom-housing/bloom/issues/2440)) ([63105bc](https://github.com/bloom-housing/bloom/commit/63105bcedfe371a4a9995e25b1e5aba67d06ad0c))
+
+
+
+
+
+## [4.0.1-alpha.39](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.38...@bloom-housing/partners@4.0.1-alpha.39) (2022-02-01)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.38](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.37...@bloom-housing/partners@4.0.1-alpha.38) (2022-02-01)
+
+
+### Features
+
+* **backend:** add publishedAt and closedAt to listing entity ([#2432](https://github.com/bloom-housing/bloom/issues/2432)) ([f3b0f86](https://github.com/bloom-housing/bloom/commit/f3b0f864a6d5d2ad3d886e828743454c3e8fca71))
+
+
+
+
+
+## [4.0.1-alpha.37](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.36...@bloom-housing/partners@4.0.1-alpha.37) (2022-02-01)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.36](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.35...@bloom-housing/partners@4.0.1-alpha.36) (2022-01-31)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.35](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.34...@bloom-housing/partners@4.0.1-alpha.35) (2022-01-31)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.34](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.33...@bloom-housing/partners@4.0.1-alpha.34) (2022-01-27)
+
+
+### Features
+
+* outdated password messaging updates ([b14e19d](https://github.com/bloom-housing/bloom/commit/b14e19d43099af2ba721d8aaaeeb2be886d05111))
+
+
+
+
+
+## [4.0.1-alpha.33](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.32...@bloom-housing/partners@4.0.1-alpha.33) (2022-01-26)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.32](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.31...@bloom-housing/partners@4.0.1-alpha.32) (2022-01-26)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.31](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.30...@bloom-housing/partners@4.0.1-alpha.31) (2022-01-26)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.30](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.29...@bloom-housing/partners@4.0.1-alpha.30) (2022-01-24)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.29](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.28...@bloom-housing/partners@4.0.1-alpha.29) (2022-01-24)
+
+
+### Bug Fixes
+
+* ami charts without all households ([#2430](https://github.com/bloom-housing/bloom/issues/2430)) ([92dfbad](https://github.com/bloom-housing/bloom/commit/92dfbad32c90d84ee1ec3a3468c084cb110aa8be))
+
+
+
+
+
+## [4.0.1-alpha.28](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.27...@bloom-housing/partners@4.0.1-alpha.28) (2022-01-21)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.27](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.26...@bloom-housing/partners@4.0.1-alpha.27) (2022-01-21)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
+## [4.0.1-alpha.26](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.25...@bloom-housing/partners@4.0.1-alpha.26) (2022-01-20)
+
+**Note:** Version bump only for package @bloom-housing/partners
+
+
+
+
+
 ## [4.0.1-alpha.25](https://github.com/bloom-housing/bloom/compare/@bloom-housing/partners@4.0.1-alpha.24...@bloom-housing/partners@4.0.1-alpha.25) (2022-01-14)
 
 **Note:** Version bump only for package @bloom-housing/partners